refactor(settings): move slack accessors and validators to settings_slack.go (PRD #1021 M3)
